Added attack and movement fix
- added attack via right click (test on golem)
- attack is inconsistent and buggy, unsure why
- added fix to stop units from vibrating when getting close to their destination

+++ Assets/Scripts/Units/Interactable.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class Interactable : MonoBehaviour
 {
     public MeshCollider myMeshCollider;
-    public int maxHealth;
-    public int currentHealth;
+    public float maxHealth;
+    public float currentHealth;
     public enum teams { neutral, allied, enemy } // team tag
     public teams team;
     public Outline outline;
